feat(auth): add login protection and owner checks for listings and reviews

+++ routes/listings.js
@@ -3,6 +3,7 @@ import wrapAsync from "../utils/wrapAsync.js";
 import ExpressError from "../utils/ExpressError.js";
 import listingSchema from "../schema.js";
 import Listing from "../models/listing.js";
+import { isLoggedIn, isOwner } from "../middleware.js";
 
 const router = express.Router();
 
@@ -29,37 +30,37 @@ const validateListing = (req, res, next) => {
   }
 };
 
-router.use((req, res, next) => {
-  res.locals.success = req.flash("success");
-  res.locals.error = req.flash("error");
-  res.locals.info = req.flash("info");
-  next();
-});
-
 router.get(
   "/",
   wrapAsync(async (req, res, next) => {
     const data = await Listing.find({});
-    res.render("listings/index.ejs", { data });
+    res.render("listings/index.ejs", { data, user: req.user });
   })
 );
 router.get(
   "/view/:id",
   wrapAsync(async (req, res, next) => {
     const { id } = req.params;
-    const item = await Listing.findById(id).populate("reviews");
+    const item = await Listing.findById(id)
+      .populate({
+        path: "reviews",
+        populate: { path: "author" },
+      })
+      .populate("owner");
     if (!item) {
       req.flash("error", "Couldn't fetch the listing !!");
       return res.redirect("/listings");
     }
-    return res.render("listings/item.ejs", { item });
+    return res.render("listings/item.ejs", { item, user: req.user });
   })
 );
-router.get("/new", (req, res, next) => {
-  res.render("listings/new.ejs");
+router.get("/new", isLoggedIn, isOwner, (req, res, next) => {
+  res.render("listings/new.ejs", { user: req.user });
 });
 router.post(
   "/new",
+  isLoggedIn,
+  isOwner,
   validateListing,
   wrapAsync(async (req, res) => {
     const { title, description, filename, url, price, location, country } =
@@ -75,6 +76,7 @@ router.post(
       location,
       country,
     };
+    newItem.owner = req.user._id;
     const item = await Listing.insertOne(newItem);
     if (item) req.flash("success", "New Listing created successfully");
     res.redirect("/listings");
@@ -82,18 +84,29 @@ router.post(
 );
 router.get(
   "/:id/edit",
+  isLoggedIn,
+  isOwner,
   wrapAsync(async (req, res, next) => {
     const { id } = req.params;
-    const item = await Listing.findById(id);
-    if (!item) {
-      req.flash("error", "Couldn't fetch the listing !!");
-      return res.redirect("/listings");
+    const listing = await Listing.findById(id).select("owner");
+    const userId = req.user._id;
+    if (listing.owner._id.equals(userId)) {
+      const item = await Listing.findById(id);
+      if (!item) {
+        req.flash("error", "Couldn't fetch the listing !!");
+        return res.redirect("/listings");
+      }
+      res.render("listings/edit.ejs", { item, user: req.user });
+    } else {
+      req.flash("error", "You must be the owner of this listing to update it.");
+      return res.redirect(`/listings/view/${id}`);
     }
-    res.render("listings/edit.ejs", { item });
   })
 );
 router.patch(
   "/:id/edit",
+  isLoggedIn,
+  isOwner,
   wrapAsync(async (req, res, next) => {
     const { id } = req.params;
     const { title, description, filename, url, price, location, country } =
@@ -118,11 +131,20 @@ router.patch(
 );
 router.delete(
   "/:id/delete",
+  isLoggedIn,
+  isOwner,
   wrapAsync(async (req, res, next) => {
     const { id } = req.params;
-    await Listing.findByIdAndDelete(id);
-    req.flash("success", "Listing deleted successfully!");
-    res.redirect("/listings");
+    const listing = await Listing.findById(id).select("owner");
+    const userId = req.user._id;
+    if (listing.owner._id.equals(userId)) {
+      await Listing.findByIdAndDelete(id);
+      req.flash("success", "Listing deleted successfully!");
+      return res.redirect("/listings");
+    } else {
+      req.flash("error", "You must be the owner of this listing to delete it.");
+      return res.redirect(`/listings/view/${id}`);
+    }
   })
 );
